Add files via upload
This is a database/collection of related tables I made from a csv file. I've included the code I used to transfer and transform the data from the CSV file, the code I used to add the constraints and an ER model of the final database. I did this using PostgreSQL and MongoDB's DBSchema application.